Fixed wrong variable type

+++ example/stepper-basic-example.c
@@ -20,7 +20,7 @@ const uint8_t stepper_pin_1A = 12;
 const uint8_t stepper_pin_1B = 13;
 const uint8_t stepper_pin_2A = 14;
 const uint8_t stepper_pin_2B = 15;
-const uint8_t stepper_steps_per_revolution = 200;
+const uint16_t stepper_steps_per_revolution = 200;
 const stepper_mode_t stepping_mode = single;
 uint8_t speed = 20;
